% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagnertimoR.R
\name{dropColsInDF}
\alias{dropColsInDF}
\title{dropColsInDF}
\usage{
dropColsInDF(df, namesVector)
}
\arguments{
\item{df}{- a data.frame}

\item{namesVector}{- the vector (c()) with the names of data.frame which should be dropped.}
}
\value{
the given data.frame without the variables (columns) in the namesVector.
}
\description{
This function removes columns from a given data.frame. The columns have to be in a vector.
}
\examples{

df <- data.frame(age = c(16,20,19), name = c("Tim", "Anne", "Mike"), weight = c(75,51,80), height = c(182,164,188))
df.new <- dropColsInDF(df, c("weight", "height"))

}
