% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagnertimoR.R
\name{formatGermanNumber}
\alias{formatGermanNumber}
\title{formatGermanNumber}
\usage{
formatGermanNumber(x)
}
\arguments{
\item{x}{- a 'german' number in the format like e.g. 123.456.789,09}
}
\value{
the converted input number with only a point as decimal delimiter. E.g. 123546789.09
}
\examples{
singlePointNumber <- formatGermanNumber("123.456.789,09")

}
