% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagnertimoR.R
\name{keepColsInDF}
\alias{keepColsInDF}
\title{keepColsInDF}
\usage{
keepColsInDF(df, namesVector)
}
\arguments{
\item{df}{- a data.frame}

\item{namesVector}{- the vector (c()) with the names of data.frame which will be kept.}
}
\value{
the given data.frame only with the variables (columns) in the namesVector.
}
\description{
This function only keeps the columns of a given data.frame from a specified namesVector.
}
\examples{

df <- data.frame(age = c(16,20,19), name = c("Tim", "Anne", "Mike"), weight = c(75,51,80), height = c(182,164,188))
df.new <- keepColsInDF(df, c("name"))

}
