% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagnertimoR.R
\name{lastDayOfMonth}
\alias{lastDayOfMonth}
\title{lastDayOfMonth}
\usage{
lastDayOfMonth(day, month, year)
}
\arguments{
\item{day}{- Specify the last week day. Sunday == 1, Monday == 2, ... Saturday == 7}

\item{month}{- Determine the month of the date}

\item{year}{- Determine the year of the date}
}
\value{
date (Date object) of the last given week day in the month and year
}
\description{
This function returns the date ("YYYY-MM-DD") of the last specified day (Sunday == 1, Monday == 2, ... Saturday == 7) in a given month (January == 1 ...)
}
\examples{

# Give me the date of the last sunday in october 2016 --> DST date
lastDayOfMonth(1,10,2016)

}
