% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagnertimoR.R
\name{left_join_fill}
\alias{left_join_fill}
\title{join_left_fill}
\usage{
left_join_fill(x, y, by, fill = 0)
}
\arguments{
\item{x}{- the left data.frame where the NAs are perserved}

\item{y}{- the right data.frame wherre NA values can occur due to the join. CAUTION valid NAs get also replaced by the fill value!}

\item{by}{- the by variable to join the data.frames}

\item{fill}{- custom fill value to replace the default NA in dplyrs join function}
}
\value{
the joined data.frame with filled 0 values
}
\description{
This function proceeds a left join (dplyr packages) and fills missing values of the 'by variable' with a given value (instead of the default NA).
Importantly, the other valid NA values are preserved. BUT only for the data.frame on the left side!
NOTE: dplyr and tidyr package are needed!
}
\examples{
dt1 <- data.table(x=c('a', 'b', 'c', 'd', 'e'), y=c(NA, 'w', NA, 'y', 'z'));
dt2 <- data.table(x=c('a', 'b', 'c'), new_col=c(1,2,3));

left_join_fill(dt1, dt2, by = x, fill = 0)

#  x    y new_col
#1 a <NA>       1
#2 b    w       2
#3 c <NA>       3
#4 d    y       0
#5 e    z       0

}
